Use GenAIButton
